This is a manual version bump by the core team,
Reasons,

The version numbering was thrown off in 2014 because of a bump from
0.99.3 --> 0.101.0, as shown in this commit
5ea1d9580a4086ecea44cef696679a7e248ff2ab.

After that, the package wasn't caught by the version bump script during
release, and escaped the 'x' version increment, and only passed the 'y'
increment.

This solution fixes the issue in the current 'master' branch, where the
author should update ONLY the 'z' number. The core team will increment
the 'y' during release in the fall. The next devel cycle will start from the
2.y.z version number.